requirements.txt
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
scikit-learn==1.4.2
python-dotenv==1.0.1
kagglehub==0.2.7
joblib==1.4.2
flask==3.0.3
gunicorn==22.0.0
google-cloud-storage==2.18.2
google-cloud-aiplatform==1.74.0
google-cloud-monitoring==2.23.1
streamlit==1.41.1
